--- a/examples/hybrid_quantum_algorithms/QAOA/QAOA_braket.ipynb
+++ b/examples/hybrid_quantum_algorithms/QAOA/QAOA_braket.ipynb
@@ -215,20 +215,9 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": null,
    "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "\u001b[33mWARNING: You are using pip version 21.2.4; however, version 23.0.1 is available.\n",
-      "You should consider upgrading via the '/Library/Developer/CommandLineTools/usr/bin/python3 -m pip install --upgrade pip' command.\u001b[0m\n",
-      "\u001b[33mWARNING: You are using pip version 21.2.4; however, version 23.0.1 is available.\n",
-      "You should consider upgrading via the '/Library/Developer/CommandLineTools/usr/bin/python3 -m pip install --upgrade pip' command.\u001b[0m\n"
-     ]
-    }
-   ],
+   "outputs": [],
    "source": [
     "# If these have already been installed, this cell can be commented out.\n",
     "!pip install pyqubo -q\n",
@@ -1018,9 +1007,18 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 35,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Version: 1.35.4.post0\n",
+      "Version: 1.11.5.post0\n"
+     ]
+    }
+   ],
    "source": [
     "# Check SDK version\n",
     "!pip show amazon-braket-sdk | grep Version\n",
@@ -1044,7 +1042,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 23,
+   "execution_count": 36,
    "metadata": {},
    "outputs": [
     {