*Description of changes:*

Reducing the runtimes for some of the most computationally heavy notebooks:

- examples/hybrid_quantum_algorithms/QAOA-QAOA_braket.ipynb: 65s -> 7s (changed the optimization method to COBYLA)

- examples/hybrid_quantum_algorithms/VQE_Transverse_Ising-VQE_Transverse_Ising_Model.ipynb: 120s -> 71s (reduced number of spins 5->4, changed the optimization method to COBYLA).
 - examples/pennylane/3_Hydrogen_Molecule_geometry_with_VQE/3_Hydrogen_Molecule_geometry_with_VQE.ipynb: 210s->70s (changed optimizer type to RMSProp and reducing the number of iterations to num_iterations=5).
 
By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.