@ykharkov 2023-02-24
@christianbmadsen Thanks for pointing this out! 
Fixed warnings in QAOA_braket.ipynb, please check.